# Speed up Util.FastCopyIntoChannel