Depends on #21358

The assets for the Tiberian Dawn HD mod are much larger than assets for the default mods, causing a lot of load time to be spent in Util.FastCopyIntoChannel.

We can provide a special case for the SpriteFrameType.Bgra32 format, which is the same format as the destination buffer. In this scenario we can just perform memory copies between the source and destination. Additionally, whilst the default mods require all their assets to get their alpha premultiplied, many of the Tiberian Dawn assets are already premultiplied. Being able to skip this step for these assets saves us having to interpret the bytes into colors and back again.

For the default mods, there isn't a noticeable timing difference. For Tiberian Dawn HD or other mods with modern assets sizes, a large speedup is achieved.

Timings for TD HD on my machine, split by assets for which `premultiplied` was true or false. When true, a bigger speedup is seen as the method doesn't have to do the premultiplication step.
| Already premultipled? | Before | After |
| -- | -- | -- |
| true | 740ms | 220ms |
| false | 340ms | 210ms |
| total | 1080ms | 430ms |